+++ src/main.tsx
@@ -2,9 +2,9 @@ import { createRoot } from "react-dom/client";
 import { createBrowserRouter, RouterProvider } from "react-router-dom";
 import './index.css'
 import App from './App.tsx'
-import Home from './pages/home/Home.tsx'
-import Login from './pages/login/Login.tsx'
-import Signup from './pages/signup/Signup.tsx'
+import Home from './pages/Home.tsx'
+import Login from './pages/Login.tsx'
+import Signup from './pages/Signup.tsx'
 
 const router = createBrowserRouter([
   {
