+++ src/pages/Login.tsx
@@ -1,8 +1,8 @@
 import React, { useState } from 'react';
 import { useForm } from 'react-hook-form';
 import { useNavigate } from 'react-router-dom';
-import { useAuthStore } from '../../store/authStore';
-import { LoginForm } from '../../types/login';
+import { useAuthStore } from '../store/authStore';
+import { LoginForm } from '../types/login';
 
 
 const Login = () => {
